Refactor textbook generator to check redirects
On a case insensitive file-system, the redirect pages could write over
the originals, changing the case of file name, and breaking the links on
a static site.  This wasn't obvious serving locally on my Mac, because
the server is also working on a case-insensitive filesystem and
therefore fetches 'numbers.html' when asked for 'Numbers.html'.  But,
when built and pushed up to - say - Github - with a case-sensitive
filesystem, the breakage appears.